final/backend/requirements.txt
streamlit
diffusers
transformers
torch
Pillow
opencv-python
accelerate
flask
flask_cors